feac: criação da rota e regra de condicao de pagamento

+++ import-control/src/main/java/com/claudio/importcontrol/controller/ProcessoController.java
@@ -60,7 +60,7 @@ public class ProcessoController {
     }
 
     @GetMapping("/quantidade/{qtd}")
-    public ResponseEntity<List<ProcessoResponseDTO>> filtrarPorQuantidade(@PathVariable Integer qtd) {
+    public ResponseEntity<List<ProcessoResponseDTO>> filtrarPorQuantidade(@PathVariable Double qtd) {
         List<ProcessoResponseDTO> lista = service.buscarMaioresQue(qtd).stream()
                 .map(ProcessoResponseDTO::new)
                 .toList();
